Timer test

+++ astrolink4pi.h
@@ -152,7 +152,7 @@ private:
 	void systemUpdate();
 	void fanControl();
 	void pwmCycle();
-	uint32_t millis();
+	long millis();
 
 	static constexpr const char *SYSTEM_TAB {"System"};
 	static constexpr const char *OUTPUTS_TAB {"Outputs"};
